Move NTimer in NTime
