+++ src/NodeEngine/Core/World.hpp
@@ -11,7 +11,7 @@
 
 #include "../Utils/Container.hpp"
 #include "../Utils/Pugixml.hpp"
-#include "../Utils/Timer.hpp"
+#include "../Utils/Time.hpp"
 #include "../Utils/Log.hpp"
 
 #include "../Application/Application.hpp"
